backend/requirements.txt
aiohttp
langchain-ollama
openai
pandas==2.2.2
pillow
python-dotenv
requests
xlsxwriter
openpyxl
pywin32
ujson
fastapi
uvicorn
pydantic
python-multipart